tvheadend: place epgdb.v3 in /tmp instead of the gone v2
Upstream switched to v3 in 2018 in their main branch effectively
making the (by default enabled) option to move the (depending on
options) heavily written to and potentially big epgdb into /tmp a
no-op since the package was moved from the 4.2 branch to main at
the end of 2021.

That said, it was broken before already as upstream uses an atomic
replace strategy for the file since the 4.1 branch (2016), which
the packaging here moved to from 4.0 in August 2021.

References: tvheadend/ae2c76c7886dbec3c2d420b734659ce97a62fc72, #17201,
 tvheadend/b23686a55323625b15d4f99fd7af55259fa21828, #16359